# apps/server/src/main/java/com/niceday/workshop/api/dto/SessionResponse.java
package com.niceday.workshop.api.dto;

public record SessionResponse(
        String id,
        String team,
        String title,
        String description,
        int runningMinutes,
        int displayOrder
) {
}