# apps/server/src/main/resources/openapi/workshop-api.yaml
openapi: 3.0.3
info:
  title: 워크샵 API
  version: 0.1.0
  description: 클라이언트/관리자 연동을 위한 읽기 API 초안
servers:
  - url: http://localhost:8080
paths:
  /api/v1/auth/login:
    post:
      summary: 로그인 인증
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthLoginRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthLoginResponse'
        '400':
          description: BAD REQUEST
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '401':
          description: UNAUTHORIZED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
  /api/v1/workshop/overview:
    get:
      summary: 워크샵 개요 통계 조회
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OverviewResponse'
  /api/v1/workshop/schedules:
    get:
      summary: 일정 목록 조회
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ScheduleItemResponse'
    post:
      summary: 일정 생성
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScheduleUpsertRequest'
      responses:
        '201':
          description: CREATED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScheduleItemResponse'
        '400':
          description: BAD REQUEST
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
  /api/v1/workshop/schedules/{id}:
    patch:
      summary: 일정 수정
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScheduleUpsertRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScheduleItemResponse'
        '400':
          description: BAD REQUEST
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '404':
          description: NOT FOUND
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
    delete:
      summary: 일정 삭제
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: NO CONTENT
        '404':
          description: NOT FOUND
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
  /api/v1/workshop/missions:
    get:
      summary: 미션 목록 조회
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MissionResponse'
    post:
      summary: 미션 생성
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MissionUpsertRequest'
      responses:
        '201':
          description: CREATED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MissionResponse'
        '400':
          description: BAD REQUEST
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
  /api/v1/workshop/missions/{id}:
    patch:
      summary: 미션 수정
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MissionUpsertRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MissionResponse'
        '400':
          description: BAD REQUEST
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '404':
          description: NOT FOUND
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
    delete:
      summary: 미션 삭제
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: NO CONTENT
        '404':
          description: NOT FOUND
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
  /api/v1/workshop/sessions:
    get:
      summary: 세션 목록 조회
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SessionResponse'
    post:
      summary: 세션 생성
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SessionUpsertRequest'
      responses:
        '201':
          description: CREATED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionResponse'
        '400':
          description: BAD REQUEST
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
  /api/v1/workshop/sessions/{id}:
    patch:
      summary: 세션 수정
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SessionUpsertRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionResponse'
        '400':
          description: BAD REQUEST
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '404':
          description: NOT FOUND
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
    delete:
      summary: 세션 삭제
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: NO CONTENT
        '404':
          description: NOT FOUND
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
  /api/v1/workshop/sessions/reorder:
    post:
      summary: 세션 카드 순서 변경
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SessionReorderRequest'
      responses:
        '204':
          description: NO CONTENT
        '400':
          description: BAD REQUEST
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
  /api/v1/workshop/sessions/{id}/questions:
    get:
      summary: 세션 질문 목록 조회
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SessionQuestionResponse'
        '404':
          description: NOT FOUND
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
    post:
      summary: 세션 질문 등록
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SessionQuestionCreateRequest'
      responses:
        '201':
          description: CREATED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionQuestionResponse'
        '400':
          description: BAD REQUEST
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '404':
          description: NOT FOUND
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
  /api/v1/workshop/sessions/{sessionId}/questions/{questionId}/answer:
    patch:
      summary: 세션 질문 답변 등록
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
        - name: questionId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SessionQuestionAnswerRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionQuestionResponse'
        '400':
          description: BAD REQUEST
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '403':
          description: FORBIDDEN
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '404':
          description: NOT FOUND
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
  /api/v1/workshop/sessions/{sessionId}/questions/{questionId}:
    delete:
      summary: 세션 질문 삭제
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
        - name: questionId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: NO CONTENT
        '400':
          description: BAD REQUEST
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '404':
          description: NOT FOUND
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
  /api/v1/workshop/users:
    get:
      summary: 사용자 목록 조회
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserResponse'
    post:
      summary: 사용자 생성
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpsertRequest'
      responses:
        '201':
          description: CREATED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: BAD REQUEST
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '409':
          description: CONFLICT
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
  /api/v1/workshop/teams:
    get:
      summary: 워크샵 팀 목록 조회
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TeamResponse'
    post:
      summary: 워크샵 팀 생성
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TeamUpsertRequest'
      responses:
        '201':
          description: CREATED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamResponse'
        '400':
          description: BAD REQUEST
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '409':
          description: CONFLICT
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
  /api/v1/workshop/teams/{id}:
    patch:
      summary: 워크샵 팀 수정
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TeamUpsertRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamResponse'
        '400':
          description: BAD REQUEST
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '404':
          description: NOT FOUND
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '409':
          description: CONFLICT
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
    delete:
      summary: 워크샵 팀 삭제
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: NO CONTENT
        '404':
          description: NOT FOUND
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
  /api/v1/workshop/users/{id}:
    patch:
      summary: 사용자 수정
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpsertRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: BAD REQUEST
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '404':
          description: NOT FOUND
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '409':
          description: CONFLICT
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
    delete:
      summary: 사용자 삭제
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: NO CONTENT
        '404':
          description: NOT FOUND
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
  /api/v1/workshop/users/{id}/password/reset:
    post:
      summary: 사용자 비밀번호 초기화
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: NO CONTENT
        '404':
          description: NOT FOUND
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
components:
  schemas:
    ApiErrorResponse:
      type: object
      properties:
        message:
          type: string
        fieldErrors:
          type: object
          additionalProperties:
            type: string
      required: [message, fieldErrors]
    OverviewResponse:
      type: object
      properties:
        activeMissions:
          type: integer
        upcomingSessions:
          type: integer
        totalUsers:
          type: integer
        totalSchedules:
          type: integer
        pendingSubmissions:
          type: integer
      required: [activeMissions, upcomingSessions, totalUsers, totalSchedules, pendingSubmissions]
    ScheduleItemResponse:
      type: object
      properties:
        id: { type: string }
        day: { type: string }
        startsAt: { type: string }
        endsAt: { type: string }
        title: { type: string }
        location: { type: string }
      required: [id, day, startsAt, endsAt, title, location]
    ScheduleUpsertRequest:
      type: object
      properties:
        day: { type: string }
        startsAt: { type: string }
        endsAt: { type: string }
        title: { type: string }
        location: { type: string }
      required: [day, startsAt, endsAt, title, location]
    MissionResponse:
      type: object
      properties:
        id: { type: string }
        title: { type: string }
        points: { type: integer }
        active: { type: boolean }
        pendingApprovals: { type: integer }
      required: [id, title, points, active, pendingApprovals]
    MissionUpsertRequest:
      type: object
      properties:
        title: { type: string }
        points: { type: integer }
        active: { type: boolean }
        pendingApprovals: { type: integer }
      required: [title, points, active, pendingApprovals]
    SessionResponse:
      type: object
      properties:
        id: { type: string }
        team: { type: string }
        title: { type: string }
        description: { type: string }
        runningMinutes: { type: integer }
        displayOrder: { type: integer }
      required: [id, team, title, description, runningMinutes, displayOrder]
    SessionUpsertRequest:
      type: object
      properties:
        team: { type: string }
        title: { type: string }
        description: { type: string }
        runningMinutes: { type: integer }
      required: [team, title, description, runningMinutes]
    SessionReorderRequest:
      type: object
      properties:
        orderedIds:
          type: array
          items:
            type: string
      required: [orderedIds]
    SessionQuestionResponse:
      type: object
      properties:
        id: { type: string }
        sessionId: { type: string }
        question: { type: string }
        answer:
          type: string
          nullable: true
        createdAt: { type: integer, format: int64 }
      required: [id, sessionId, question, createdAt]
    SessionQuestionCreateRequest:
      type: object
      properties:
        question: { type: string }
      required: [question]
    SessionQuestionAnswerRequest:
      type: object
      properties:
        answer: { type: string }
      required: [answer]
    UserResponse:
      type: object
      properties:
        id: { type: string }
        username: { type: string }
        name: { type: string }
        team: { type: string }
        workshopTeamId:
          type: string
          nullable: true
        workshopTeamName:
          type: string
          nullable: true
        department: { type: string }
        role: { type: string }
      required: [id, username, name, team, department, role]
    UserUpsertRequest:
      type: object
      properties:
        username: { type: string }
        name: { type: string }
        team: { type: string }
        workshopTeamId:
          type: string
          nullable: true
        department:
          type: string
          nullable: true
        role: { type: string }
      required: [username, name, team, role]
    TeamResponse:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
      required: [id, name]
    TeamUpsertRequest:
      type: object
      properties:
        name: { type: string }
      required: [name]
    AuthLoginRequest:
      type: object
      properties:
        username: { type: string }
        password: { type: string }
      required: [username, password]
    AuthLoginResponse:
      type: object
      properties:
        role: { type: string }
      required: [role]